Add documentation, fix schema
This adds language documentation, and fixes a bug in the schema
that effectively rejected footnote references that were referring to
footnotes in the wrong scope. The schema now accurately reflects the
fact that footnote references can only reference footnotes in the
containing section, as was intended all along.

Affects: https://github.com/io7m/xstructural/issues/13